Adding dates to array commit
Created CaptureData function to store booked dates with .push().

CaptureData originally duplicate dates, and the web app should only allow 1 appointment booking per day.

So I added a helper method isInArray to check for duplicate dates, and only add a date if it wasn't a duplicate.

This ensures no duplicates are in the default array of dates.

Got onclick on delete button to call 2 functions. This allowed me to remove the appointment from 1) the frontend list and 2) the booked_dates array.

Next, I need to run a check against all current booked_dates when the user enters a new date.

If it's a duplicate, display an error.